+++ scilab/modules/graphics/sci_gateway/c/sci_grayplot.c
@@ -48,7 +48,7 @@ int sci_grayplot( char *fname, unsigned long fname_len )
 
     if (Rhs <= 0)
     {
-        sci_demo(fname, fname_len);
+        sci_demo(fname, (int)fname_len);
         return 0;
     }
     CheckRhs(3, 7);
