+++ scilab/modules/graphics/sci_gateway/c/sci_xset.c
@@ -55,7 +55,7 @@ int sci_xset( char *fname, unsigned long fname_len )
 
     if (Rhs <= 0)
     {
-        sci_demo(fname, fname_len);
+        sci_demo(fname, (int)fname_len);
         return 0;
     }
